backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
tensorflow==2.17.0
transformers==4.38.2
tf-keras
numpy
Pillow
scipy